Abstract out handlers from routers
