+++ server/views/email/forgot-password-email.hbs
