+++ server/views/email/reset-password-email.hbs
