+++ server/views/forgot-password.hbs
@@ -1,4 +1,4 @@
-<form action="/forgot_password" method="post">
+<form action="/forgot-password" method="post">
     <div class="form-group">
         <label>Please type in your email address</label>
         <input type="email" class="form-control" name="email">
